add memory proj to perf analysis
