+++ data/Llama2-13B/IntelGaudi2B_overall_projection.csv
@@ -18,8 +18,8 @@ Llama2-13B,IntelGaudi2B,1,1,BF16,1024,512,16,256,815.05,19.87,21.56,20.72,22.27,
 Llama2-13B,IntelGaudi2B,1,1,BF16,1024,512,32,256,1619.79,28.38,31.76,30.07,33.17,964.73
 Llama2-13B,IntelGaudi2B,1,1,BF16,1024,512,64,256,3229.27,45.39,52.16,48.77,54.98,1164.06
 Llama2-13B,IntelGaudi2B,1,1,BF16,1024,512,128,256,6448.22,79.43,92.95,86.17,98.6,1298.17
-Llama2-13B,IntelGaudi2B,1,1,BF16,1024,512,256,256,12886.13,152.75,179.8,166.24,191.09,1339.68
-Llama2-13B,IntelGaudi2B,1,1,BF16,1024,512,512,256,25761.95,302.99,357.1,329.99,379.67,1348.54
+Llama2-13B,IntelGaudi2B,1,1,BF16,1024,512,256,256,12886.13,152.75,179.8,166.24,191.09,OOM
+Llama2-13B,IntelGaudi2B,1,1,BF16,1024,512,512,256,25761.95,302.99,357.1,329.99,379.67,OOM
 ,,,,,,,,,,,,,,
 Llama2-13B,IntelGaudi2B,1,1,BF16,2048,512,1,256,121.82,12.32,12.43,12.37,12.59,79.43
 Llama2-13B,IntelGaudi2B,1,1,BF16,2048,512,2,256,233.32,13.28,13.49,13.38,13.81,144.82
@@ -28,7 +28,7 @@ Llama2-13B,IntelGaudi2B,1,1,BF16,2048,512,8,256,902.35,19.0,19.85,19.42,21.15,37
 Llama2-13B,IntelGaudi2B,1,1,BF16,2048,512,16,256,1794.38,26.64,28.33,27.48,30.93,517.3
 Llama2-13B,IntelGaudi2B,1,1,BF16,2048,512,32,256,3578.45,41.91,45.29,43.59,50.5,633.66
 Llama2-13B,IntelGaudi2B,1,1,BF16,2048,512,64,256,7146.59,72.45,79.21,75.82,89.63,714.05
-Llama2-13B,IntelGaudi2B,1,1,BF16,2048,512,128,256,14282.87,133.53,147.05,140.28,167.9,762.36
-Llama2-13B,IntelGaudi2B,1,1,BF16,2048,512,256,256,28555.43,260.95,288.0,274.45,329.69,776.49
-Llama2-13B,IntelGaudi2B,1,1,BF16,2048,512,512,256,57100.54,519.41,573.51,546.41,656.86,779.47
+Llama2-13B,IntelGaudi2B,1,1,BF16,2048,512,128,256,14282.87,133.53,147.05,140.28,167.9,OOM
+Llama2-13B,IntelGaudi2B,1,1,BF16,2048,512,256,256,28555.43,260.95,288.0,274.45,329.69,OOM
+Llama2-13B,IntelGaudi2B,1,1,BF16,2048,512,512,256,57100.54,519.41,573.51,546.41,656.86,OOM
 ,,,,,,,,,,,,,,
