+++ data/Qwen-14B/IntelGaudi2B_overall_projection.csv
@@ -18,8 +18,8 @@ Qwen-14B,IntelGaudi2B,1,1,BF16,1024,512,16,256,810.73,19.8,21.49,20.65,22.19,721
 Qwen-14B,IntelGaudi2B,1,1,BF16,1024,512,32,256,1611.21,28.31,31.69,30.0,33.09,967.06
 Qwen-14B,IntelGaudi2B,1,1,BF16,1024,512,64,256,3212.16,45.32,52.09,48.7,54.88,1166.18
 Qwen-14B,IntelGaudi2B,1,1,BF16,1024,512,128,256,6414.07,79.35,92.88,86.1,98.46,1300.02
-Qwen-14B,IntelGaudi2B,1,1,BF16,1024,512,256,256,12817.9,152.64,179.69,166.14,190.85,1341.37
-Qwen-14B,IntelGaudi2B,1,1,BF16,1024,512,512,256,25625.54,302.8,356.9,329.8,379.2,1350.21
+Qwen-14B,IntelGaudi2B,1,1,BF16,1024,512,256,256,12817.9,152.64,179.69,166.14,190.85,OOM
+Qwen-14B,IntelGaudi2B,1,1,BF16,1024,512,512,256,25625.54,302.8,356.9,329.8,379.2,OOM
 ,,,,,,,,,,,,,,
 Qwen-14B,IntelGaudi2B,1,1,BF16,2048,512,1,256,121.22,12.25,12.36,12.3,12.52,79.87
 Qwen-14B,IntelGaudi2B,1,1,BF16,2048,512,2,256,232.19,13.2,13.42,13.31,13.74,145.56
@@ -28,7 +28,7 @@ Qwen-14B,IntelGaudi2B,1,1,BF16,2048,512,8,256,898.03,18.93,19.78,19.35,21.07,379
 Qwen-14B,IntelGaudi2B,1,1,BF16,2048,512,16,256,1785.8,26.57,28.26,27.41,30.84,518.81
 Qwen-14B,IntelGaudi2B,1,1,BF16,2048,512,32,256,3561.35,41.84,45.22,43.52,50.39,635.05
 Qwen-14B,IntelGaudi2B,1,1,BF16,2048,512,64,256,7112.44,72.38,79.14,75.75,89.49,715.16
-Qwen-14B,IntelGaudi2B,1,1,BF16,2048,512,128,256,14214.63,133.46,146.98,140.21,167.69,763.31
-Qwen-14B,IntelGaudi2B,1,1,BF16,2048,512,256,256,28419.01,260.85,287.9,274.34,329.31,777.38
-Qwen-14B,IntelGaudi2B,1,1,BF16,2048,512,512,256,56827.78,519.21,573.31,546.21,656.13,780.33
+Qwen-14B,IntelGaudi2B,1,1,BF16,2048,512,128,256,14214.63,133.46,146.98,140.21,167.69,OOM
+Qwen-14B,IntelGaudi2B,1,1,BF16,2048,512,256,256,28419.01,260.85,287.9,274.34,329.31,OOM
+Qwen-14B,IntelGaudi2B,1,1,BF16,2048,512,512,256,56827.78,519.21,573.31,546.21,656.13,OOM
 ,,,,,,,,,,,,,,
