+++ data/Qwen-7B/IntelGaudi2B_overall_projection.csv
@@ -19,7 +19,7 @@ Qwen-7B,IntelGaudi2B,1,1,BF16,1024,512,32,256,865.59,16.69,18.85,17.77,19.43,164
 Qwen-7B,IntelGaudi2B,1,1,BF16,1024,512,64,256,1725.91,27.58,31.91,29.74,33.05,1936.46
 Qwen-7B,IntelGaudi2B,1,1,BF16,1024,512,128,256,3446.56,49.36,58.01,53.68,60.3,2122.72
 Qwen-7B,IntelGaudi2B,1,1,BF16,1024,512,256,256,6887.86,95.6,112.91,104.24,117.49,2178.91
-Qwen-7B,IntelGaudi2B,1,1,BF16,1024,512,512,256,13770.46,189.93,224.55,207.21,233.7,2190.84
+Qwen-7B,IntelGaudi2B,1,1,BF16,1024,512,512,256,13770.46,189.93,224.55,207.21,233.7,OOM
 ,,,,,,,,,,,,,,
 Qwen-7B,IntelGaudi2B,1,1,BF16,2048,512,1,256,66.02,6.41,6.48,6.45,6.56,152.44
 Qwen-7B,IntelGaudi2B,1,1,BF16,2048,512,2,256,126.77,7.02,7.16,7.09,7.32,273.22
@@ -28,7 +28,7 @@ Qwen-7B,IntelGaudi2B,1,1,BF16,2048,512,8,256,491.3,10.69,11.23,10.96,11.9,672.27
 Qwen-7B,IntelGaudi2B,1,1,BF16,2048,512,16,256,977.33,15.57,16.66,16.11,17.99,889.38
 Qwen-7B,IntelGaudi2B,1,1,BF16,2048,512,32,256,1949.39,25.35,27.51,26.43,30.18,1060.3
 Qwen-7B,IntelGaudi2B,1,1,BF16,2048,512,64,256,3893.52,44.89,49.22,47.05,54.57,1172.81
-Qwen-7B,IntelGaudi2B,1,1,BF16,2048,512,128,256,7781.77,83.98,92.64,88.3,103.33,1238.75
-Qwen-7B,IntelGaudi2B,1,1,BF16,2048,512,256,256,15558.28,164.85,182.17,173.49,203.54,1257.74
-Qwen-7B,IntelGaudi2B,1,1,BF16,2048,512,512,256,31111.3,328.43,363.06,345.71,405.8,1261.71
+Qwen-7B,IntelGaudi2B,1,1,BF16,2048,512,128,256,7781.77,83.98,92.64,88.3,103.33,OOM
+Qwen-7B,IntelGaudi2B,1,1,BF16,2048,512,256,256,15558.28,164.85,182.17,173.49,203.54,OOM
+Qwen-7B,IntelGaudi2B,1,1,BF16,2048,512,512,256,31111.3,328.43,363.06,345.71,405.8,OOM
 ,,,,,,,,,,,,,,
