+++ data/Qwen-7B/IntelGaudi2B_pp1_tp1_overall_projection_in_detatil.csv
@@ -19,7 +19,7 @@ BF16,1024,512,32,4096,32,32,11008,True,1,32,256,865.59,16.69,18.85,17.77,19.43,1
 BF16,1024,512,64,4096,32,32,11008,True,1,32,256,1725.91,27.58,31.91,29.74,33.05,1936.46
 BF16,1024,512,128,4096,32,32,11008,True,1,32,256,3446.56,49.36,58.01,53.68,60.3,2122.72
 BF16,1024,512,256,4096,32,32,11008,True,1,32,256,6887.86,95.6,112.91,104.24,117.49,2178.91
-BF16,1024,512,512,4096,32,32,11008,True,1,32,256,13770.46,189.93,224.55,207.21,233.7,2190.84
+BF16,1024,512,512,4096,32,32,11008,True,1,32,256,13770.46,189.93,224.55,207.21,233.7,OOM
 ,,,,,,,,,,,,,,,,,
 BF16,2048,512,1,4096,32,32,11008,True,1,32,256,66.02,6.41,6.48,6.45,6.56,152.44
 BF16,2048,512,2,4096,32,32,11008,True,1,32,256,126.77,7.02,7.16,7.09,7.32,273.22
@@ -28,7 +28,7 @@ BF16,2048,512,8,4096,32,32,11008,True,1,32,256,491.3,10.69,11.23,10.96,11.9,672.
 BF16,2048,512,16,4096,32,32,11008,True,1,32,256,977.33,15.57,16.66,16.11,17.99,889.38
 BF16,2048,512,32,4096,32,32,11008,True,1,32,256,1949.39,25.35,27.51,26.43,30.18,1060.3
 BF16,2048,512,64,4096,32,32,11008,True,1,32,256,3893.52,44.89,49.22,47.05,54.57,1172.81
-BF16,2048,512,128,4096,32,32,11008,True,1,32,256,7781.77,83.98,92.64,88.3,103.33,1238.75
-BF16,2048,512,256,4096,32,32,11008,True,1,32,256,15558.28,164.85,182.17,173.49,203.54,1257.74
-BF16,2048,512,512,4096,32,32,11008,True,1,32,256,31111.3,328.43,363.06,345.71,405.8,1261.71
+BF16,2048,512,128,4096,32,32,11008,True,1,32,256,7781.77,83.98,92.64,88.3,103.33,OOM
+BF16,2048,512,256,4096,32,32,11008,True,1,32,256,15558.28,164.85,182.17,173.49,203.54,OOM
+BF16,2048,512,512,4096,32,32,11008,True,1,32,256,31111.3,328.43,363.06,345.71,405.8,OOM
 ,,,,,,,,,,,,,,,,,
